+++ migrations/2024_01_31_000002_add_lottery_participants_table.php
@@ -3,23 +3,35 @@
 use Flarum\Database\Migration;
 use Illuminate\Database\Schema\Blueprint;
 
-return Migration::createTable('lottery_participants', function (Blueprint $table) {
-    $table->increments('id');
-    $table->unsignedInteger('lottery_id');
-    $table->unsignedTinyInteger('status')->comment('0 未中奖 1中奖');
-    $table->unsignedInteger('user_id')->nullable();
-    $table->timestamps();
+return [
+    'up' => function (Builder $schema) {
+        if ($schema->hasTable('lottery_participants')) {
+            return;
+        }
 
-    $table->index('lottery_id');
-    $table->index('user_id');
+        $schema->create('lottery_participants', function (Blueprint $table) {
+            $table->increments('id');
+            $table->unsignedInteger('lottery_id');
+            $table->unsignedTinyInteger('status')->comment('0 未中奖 1中奖');
+            $table->unsignedInteger('user_id')->nullable();
+            $table->timestamps();
 
-    $table->foreign('lottery_id')
-        ->references('id')
-        ->on('lotteries')
-        ->onDelete('cascade');
+            $table->index('lottery_id');
+            $table->index('user_id');
 
-    $table->foreign('user_id')
-        ->references('id')
-        ->on('users')
-        ->onDelete('set null');
-});
+            $table->foreign('lottery_id')
+                ->references('id')
+                ->on('lotteries')
+                ->onDelete('cascade');
+
+            $table->foreign('user_id')
+                ->references('id')
+                ->on('users')
+                ->onDelete('set null');
+        });
+
+    },
+    'down' => function (Builder $schema) {
+
+    },
+];
